requirements.txt
pynput
pyaudio
pyttsx3
assemblyai
youchat-py
wave
playsound==1.2.2
setuptools-rust
git+https://github.com/openai/whisper.git
